flux_rustc_bridge
Module ty

A simplified version of rust types.

Modules

subst 🔒

Structs

AdtDef

AdtDefData

AliasTy

Binder

BoundRegion

Clause

ClosureArgs

ClosureArgsParts

Const

CoroutineArgs

CoroutineArgsParts

EarlyBinder

ExistentialProjection

ExistentialTraitRef

FieldDef

FnSig

GenericParamDef

GenericPredicates

Generics

LateParamRegion

OutlivesPredicate

ProjectionPredicate

TraitPredicate

TraitRef

Ty

TyS 🔒

UnevaluatedConst

VariantDef

BoundVar

ConstVid

A const variable ID.

DebruijnIndex

A De Bruijn index is a standard means of representing regions (and perhaps later types) in a higher-ranked setting. In particular, imagine a type like this:

EarlyParamRegion

FieldIdx

The source-order index of a field in a variant.

ParamTy

RegionVid

A region variable ID.

ScalarInt

The raw bytes of a simple value.

VariantIdx

The source-order index of a variant in a type.

Enums

AliasKind

BoundVariableKind

ClauseKind

ConstKind

ExistentialPredicate

GenericArg

GenericParamDefKind

Region

TyKind

ValTree

BoundRegionKind

FloatTy

InferConst

An inference variable for a const, for use in const generics.

IntTy

Mutability

UintTy

Constants

FIRST_VARIANT

Equivalent to VariantIdx(0).

Traits

GenericArgsExt

Functions

region_to_string

Type Aliases

GenericArgs

List

PolyExistentialPredicate

PolyFnSig

PolyTraitRef

TypeOutlivesPredicate