Crate flux_rustc_bridge

This crate contains simplified versions of some structures in rustc. The definitions in this module can be understood as the current supported subset of rust. As we implement more features we should be able to work directly on rustc’s structures.

Modules

const_eval

lowering

mir

A simplified version of rust mir.

ty

A simplified version of rust types.

Statics

DEFAULT_LOCALE_RESOURCE

Raw content of Fluent resource for this crate, generated by fluent_messages macro, imported by rustc_driver to include all crates’ resources in one bundle.

Traits

ToRustc

Functions

def_id_to_string